Rename data writer file
