+++ dds/src/implementation/dds_impl/dcps_service.rs
@@ -61,7 +61,7 @@ use super::{
     message_receiver::MessageReceiver,
     participant_discovery::ParticipantDiscovery,
     status_listener::StatusListener,
-    user_defined_data_writer::DdsDataWriter,
+    dds_data_writer::DdsDataWriter,
     user_defined_subscriber::UserDefinedSubscriber,
 };
